First Dc Bot

+++ drum_bot.py
@@ -127,7 +127,7 @@ async def react(ctx, chat:Optional[discord.Message], emoji):
         await message.add_reaction(emoji)
 
 @bot.command()
-async def solve(ctx, num1:int, operation, num2:int):
+async def solve(ctx, num1:float, operation, num2:float):
     try:
         if operation == "+":
             await ctx.send(f"{ctx.author.mention}  {num1} + {num2} = {num1 + num2}")
